+++ dicom/client.go
@@ -1,10 +1,11 @@
-// Package has provides support for HSDP Appstreaming service
-package has
+// Package dicom provides support for HSDP DICOM services
+package dicom
 
 import (
 	"bytes"
 	"encoding/json"
 	"fmt"
+	"github.com/google/go-querystring/query"
 	"github.com/philips-software/go-hsdp-api/internal"
 	"io"
 	"io/ioutil"
@@ -12,15 +13,16 @@ import (
 	"net/http/httputil"
 	"net/url"
 	"os"
-	"sort"
 	"strings"
 
-	"github.com/google/go-querystring/query"
+	"github.com/google/fhir/go/jsonformat"
+
 	"github.com/philips-software/go-hsdp-api/iam"
 )
 
 const (
-	userAgent = "go-hsdp-api/has/" + internal.LibraryVersion
+	userAgent  = "go-hsdp-api/dicom/" + internal.LibraryVersion
+	APIVersion = "1"
 )
 
 // OptionFunc is the function signature function for options
@@ -28,46 +30,44 @@ type OptionFunc func(*http.Request) error
 
 // Config contains the configuration of a client
 type Config struct {
-	HASURL   string
-	OrgID    string
-	Debug    bool
-	DebugLog string
+	Region         string
+	Environment    string
+	OrganizationID string
+	DICOMConfigURL string
+	Type           string
+	TimeZone       string
+	DebugLog       string
 }
 
-// A Client manages communication with HSDP IAM API
+// A Client manages communication with HSDP DICOM API
 type Client struct {
-	// HTTP client used to communicate with the API.
+	// HTTP client used to communicate with IAM API
 	iamClient *iam.Client
 
 	config *Config
 
-	baseHASURL *url.URL
+	dicomStoreURL *url.URL
 
-	// User agent used when communicating with the HSDP IAM API.
+	// User agent used when communicating with the HSDP DICOM API.
 	UserAgent string
 
 	debugFile *os.File
 
-	Resources *ResourcesService
-	Sessions  *SessionsService
-	Images    *ImagesService
+	Config *ConfigService
 }
 
-// NewClient returns a new HSDP HAS API client. If a nil httpClient is
-// provided, http.DefaultClient will be used. A configured IAM client must be provided
-// as well
+// NewClient returns a new HSDP DICOM API client. Configured console and IAM clients
+// must be provided as the underlying API requires tokens from respective services
 func NewClient(iamClient *iam.Client, config *Config) (*Client, error) {
 	return newClient(iamClient, config)
 }
 
 func newClient(iamClient *iam.Client, config *Config) (*Client, error) {
 	c := &Client{iamClient: iamClient, config: config, UserAgent: userAgent}
-	if err := c.SetBaseHASURL(c.config.HASURL); err != nil {
+	dicomStore := config.DICOMConfigURL + "/store/dicom/"
+
+	if err := c.SetDICOMStoreURL(dicomStore); err != nil {
 		return nil, err
-	}
-	if config.OrgID == "" || !iamClient.HasPermissions(config.OrgID,
-		"HAS_SESSION.ALL", "HAS_RESOURCE.ALL") {
-		return nil, ErrMissingHASPermissions
 	}
 	if config.DebugLog != "" {
 		var err error
@@ -76,10 +76,17 @@ func newClient(iamClient *iam.Client, config *Config) (*Client, error) {
 			c.debugFile = nil
 		}
 	}
+	ma, err := jsonformat.NewMarshaller(false, "", "", jsonformat.STU3)
+	if err != nil {
+		return nil, fmt.Errorf("dicom.NewClient create FHIR STU3 marshaller: %w", err)
+	}
+	um, err := jsonformat.NewUnmarshaller(config.TimeZone, jsonformat.STU3)
+	if err != nil {
+		return nil, fmt.Errorf("dicom.NewClient create FHIR STU3 unmarshaller (timezone=[%s]): %w", config.TimeZone, err)
+	}
 
-	c.Resources = &ResourcesService{client: c, orgID: config.OrgID}
-	c.Sessions = &SessionsService{client: c, orgID: config.OrgID}
-	c.Images = &ImagesService{client: c, orgID: config.OrgID}
+	c.Config = &ConfigService{client: c, ma: ma, um: um, profile: "production"}
+
 	return c, nil
 }
 
@@ -91,39 +98,84 @@ func (c *Client) Close() {
 	}
 }
 
-// SetBaseHASURL sets the base URL for API requests to a custom endpoint. urlStr
+func (o *Client) GetSTOWURL() string {
+	return o.replaceConfigWith("stow")
+}
+
+func (o *Client) GetWADOURL() string {
+	return o.replaceConfigWith("wado")
+}
+
+func (o *Client) GetQIDOURL() string {
+	return o.replaceConfigWith("qido")
+}
+
+func (o *Client) replaceConfigWith(svc string) string {
+	return strings.Replace(o.config.DICOMConfigURL, "config", svc, 1)
+}
+
+// GetDICOMStoreURL returns the base FHIR Store base URL as configured
+func (c *Client) GetDICOMStoreURL() string {
+	if c.dicomStoreURL == nil {
+		return ""
+	}
+	return c.dicomStoreURL.String()
+}
+
+// GetEndpointURL returns the FHIR Store Endpoint URL as configured
+func (c *Client) GetEndpointURL() string {
+	return c.GetDICOMStoreURL() + c.config.OrganizationID
+}
+
+// SetDICOMStoreURL sets the FHIR store URL for API requests to a custom endpoint. urlStr
 // should always be specified with a trailing slash.
-func (c *Client) SetBaseHASURL(urlStr string) error {
+func (c *Client) SetDICOMStoreURL(urlStr string) error {
 	if urlStr == "" {
-		return ErrBaseHASCannotBeEmpty
+		return ErrDICOMURLCannotBeEmpty
 	}
 	// Make sure the given URL end with a slash
 	if !strings.HasSuffix(urlStr, "/") {
 		urlStr += "/"
 	}
-
 	var err error
-	c.baseHASURL, err = url.Parse(urlStr)
+	c.dicomStoreURL, err = url.Parse(urlStr)
 	return err
 }
 
-// newHASRequest creates an new HAS API request. A relative URL path can be provided in
+// SetEndpointURL sets the FHIR endpoint URL for API requests to a custom endpoint. urlStr
+// should always be specified with a trailing slash.
+func (c *Client) SetEndpointURL(urlStr string) error {
+	if urlStr == "" {
+		return ErrDICOMURLCannotBeEmpty
+	}
+	// Make sure the given URL end with a slash
+	if !strings.HasSuffix(urlStr, "/") {
+		urlStr += "/"
+	}
+	var err error
+	c.dicomStoreURL, err = url.Parse(urlStr)
+	if err != nil {
+		return err
+	}
+	parts := strings.Split(c.dicomStoreURL.Path, "/")
+	if len(parts) == 0 {
+		return ErrDICOMURLCannotBeEmpty
+	}
+	c.config.OrganizationID = parts[len(parts)-1]
+	newParts := parts[:len(parts)-1]
+	c.dicomStoreURL.Path = strings.Join(newParts, "/")
+	return nil
+}
+
+// newDICOMRequest creates an new DICOM Service API request. A relative URL path can be provided in
 // urlStr, in which case it is resolved relative to the base URL of the Client.
 // Relative URL paths should always be specified without a preceding slash. If
 // specified, the value pointed to by body is JSON encoded and included as the
 // request body.
-func (c *Client) newHASRequest(method, path string, opt interface{}, options []OptionFunc) (*http.Request, error) {
-	u := *c.baseHASURL
+func (c *Client) newDICOMRequest(method, path string, bodyBytes []byte, opt interface{}, options ...OptionFunc) (*http.Request, error) {
+	u := *c.dicomStoreURL
 	// Set the encoded opaque data
-	u.Opaque = c.baseHASURL.Path + path
-
-	if opt != nil {
-		q, err := query.Values(opt)
-		if err != nil {
-			return nil, err
-		}
-		u.RawQuery = q.Encode()
-	}
+	u.Opaque = c.dicomStoreURL.Path + path
 
 	req := &http.Request{
 		Method:     method,
@@ -133,6 +185,13 @@ func (c *Client) newHASRequest(method, path string, opt interface{}, options []O
 		ProtoMinor: 1,
 		Header:     make(http.Header),
 		Host:       u.Host,
+	}
+	if opt != nil {
+		q, err := query.Values(opt)
+		if err != nil {
+			return nil, err
+		}
+		u.RawQuery = q.Encode()
 	}
 
 	for _, fn := range options {
@@ -145,21 +204,18 @@ func (c *Client) newHASRequest(method, path string, opt interface{}, options []O
 		}
 	}
 
-	if method == "POST" || method == "PUT" {
-		bodyBytes, err := json.Marshal(opt)
-		if err != nil {
-			return nil, err
-		}
+	if method == "POST" || method == "PUT" || method == "PATCH" {
 		bodyReader := bytes.NewReader(bodyBytes)
-
-		u.RawQuery = ""
 		req.Body = ioutil.NopCloser(bodyReader)
 		req.ContentLength = int64(bodyReader.Len())
-		req.Header.Set("Content-Type", "application/json")
 	}
 
 	req.Header.Set("Accept", "application/json")
 	req.Header.Set("Authorization", "Bearer "+c.iamClient.Token())
+	req.Header.Set("API-Version", APIVersion)
+	if c.config.OrganizationID != "" {
+		req.Header.Set("OrganizationID", c.config.OrganizationID)
+	}
 
 	if c.UserAgent != "" {
 		req.Header.Set("User-Agent", c.UserAgent)
@@ -197,7 +253,6 @@ func (c *Client) do(req *http.Request, v interface{}) (*Response, error) {
 	if err != nil {
 		return nil, err
 	}
-	defer resp.Body.Close()
 
 	response := newResponse(resp)
 
@@ -209,6 +264,7 @@ func (c *Client) do(req *http.Request, v interface{}) (*Response, error) {
 	}
 
 	if v != nil {
+		defer resp.Body.Close() // Only close if we plan to read it
 		if w, ok := v.(io.Writer); ok {
 			_, err = io.Copy(w, resp.Body)
 		} else {
@@ -219,64 +275,11 @@ func (c *Client) do(req *http.Request, v interface{}) (*Response, error) {
 	return response, err
 }
 
-// ErrorResponse represents an IAM errors response
-// containing a code and a human readable message
-type ErrorResponse struct {
-	Response *http.Response `json:"-"`
-	Code     string         `json:"responseCode"`
-	Message  string         `json:"responseMessage"`
-}
-
-func (e *ErrorResponse) Error() string {
-	path, _ := url.QueryUnescape(e.Response.Request.URL.Opaque)
-	u := fmt.Sprintf("%s://%s%s", e.Response.Request.URL.Scheme, e.Response.Request.URL.Host, path)
-	return fmt.Sprintf("%s %s: %d %s", e.Response.Request.Method, u, e.Response.StatusCode, e.Message)
-}
-
+// checkResponse checks the API response for errors, and returns them if present.
 func checkResponse(r *http.Response) error {
 	switch r.StatusCode {
 	case 200, 201, 202, 204, 304:
 		return nil
 	}
-
-	errorResponse := &ErrorResponse{Response: r}
-	data, err := ioutil.ReadAll(r.Body)
-	if err == nil && data != nil {
-		var raw interface{}
-		if err := json.Unmarshal(data, &raw); err != nil {
-			errorResponse.Message = "failed to parse unknown error format"
-		}
-
-		errorResponse.Message = parseError(raw)
-	}
-
-	return errorResponse
-}
-
-func parseError(raw interface{}) string {
-	switch raw := raw.(type) {
-	case string:
-		return raw
-
-	case []interface{}:
-		var errs []string
-		for _, v := range raw {
-			errs = append(errs, parseError(v))
-		}
-		return fmt.Sprintf("[%s]", strings.Join(errs, ", "))
-
-	case map[string]interface{}:
-		var errs []string
-		for k, v := range raw {
-			errs = append(errs, fmt.Sprintf("{%s: %s}", k, parseError(v)))
-		}
-		sort.Strings(errs)
-		return strings.Join(errs, ", ")
-	case float64:
-		return fmt.Sprintf("%d", int64(raw))
-	case int64:
-		return fmt.Sprintf("%d", raw)
-	default:
-		return fmt.Sprintf("failed to parse unexpected error type: %T", raw)
-	}
+	return ErrNonHttp20xResponse
 }