+++ iam/client.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/philips-software/go-hsdp-api/internal"
 	"io"
 	"io/ioutil"
 	"net/http"
@@ -28,8 +29,7 @@ type tokenType int
 type ContextKey string
 
 const (
-	libraryVersion  = "0.29.0"
-	userAgent       = "go-hsdp-api/iam/" + libraryVersion
+	userAgent       = "go-hsdp-api/iam/" + internal.LibraryVersion
 	loginAPIVersion = "2"
 )