+++ tdr/client.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/philips-software/go-hsdp-api/internal"
 	"io"
 	"io/ioutil"
 	"net/http"
@@ -20,8 +21,7 @@ import (
 )
 
 const (
-	libraryVersion = "0.29.0"
-	userAgent      = "go-hsdp-api/tdr/" + libraryVersion
+	userAgent = "go-hsdp-api/tdr/" + internal.LibraryVersion
 )
 
 // OptionFunc is the function signature function for options
@@ -83,7 +83,7 @@ func newClient(iamClient *iam.Client, config *Config) (*Client, error) {
 // Close releases allocated resources of clients
 func (c *Client) Close() {
 	if c.debugFile != nil {
-		c.debugFile.Close()
+		_ = c.debugFile.Close()
 		c.debugFile = nil
 	}
 }